requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.4
Werkzeug==2.3.6
pycryptodome==3.18.0
python-socketio==5.8.0
eventlet==0.33.3